requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.15
djangorestframework-simplejwt>=5.3
django-environ>=0.11
psycopg2-binary>=2.9
django-cors-headers>=4.3
channels>=4.1
channels-redis>=4.2
django-filter>=24.1
cryptography>=42.0
drf-spectacular>=0.27.2
uvicorn[standard]>=0.29
